# Improve Update Gamelist logic and performance