This PR improves the Gamelist Update logic on save in a few ways:
- Iteration logic: instead of iterating the entire XML for the system (larger data set) for each changed game (smaller data set), it iterates the entire XML only once (or twice, if folders can get updated somehow as well? I kept that logic though I'm not sure if it ever runs);
- Less expensive transformations: instead of relying on a path transformation that stats the file path every time - for each entry in the gamelist and then for the path in the in-memory FileData - we now just apply the same transformation we use when loading the path to the FileData structure in memory which is less expensive and we're effectively just using them to match whether that FileData corresponds to that XML element. This makes it more performant in slower media, especially remote drives.
- More robust XML handling: the previous XML element deletion would delete the iterator we were using, making subsequent behavior unreliable, resulting on occasion in missing elements for deletion. Gamelists were keeping more than one entry of the same game over time, when that behavior occurred. 

For a Gamelist with 1400 entries, here are the results on external storage:

Old version:
- 3 games changed.
- Time per gamelist entry 30ms-100ms.
- Exiting ES took ~455 seconds/over 7 mins (~100ms per entry, with added leeway for other exiting processing).

Old version without stat64-ing files, using cheaper transformation:
- 3 games changed.
- Time per gamelist entry <4ms.
- Exiting ES took 14 seconds.

New version, without stat64-ing files, using cheaper transformation and improved iteration logic:
- 3 games changed.
- Time per gamelist entry <1ms.
- Exiting ES took 8 seconds.

No observable change in behavior as far as I can tell. If anyone wants to take a look at it, it'd be appreciated, even if for sanity checking the code.

Thanks.